requirements.txt
oci
gradio
python-dotenv
